Changes to ->HitBy[] and misc parser tweaks. Changelog: npc->HitBy[] and Link->HitBy[] are now a size of 10 each. For npc->HitBy[], bot the screen index of the weapon that hits an npc, and its 'script_UID' are simultaneously stored.
**I tried to do this with Link, but I was greeted with an instability and thus,
I reverted that change for the present.

Obviously, these use radicallly different value ranges, so they are carefully
handled and vbounded as needed in ffscript.cpp. Both arrays are untyped, as the
data exchange requires it.

I also fixed the type of npc, npcdata, and itemdata->Attributes[]. Obviously,
like with ->Misc[], this is untyped, to permoit various forms of data communication.

Last, version ID bump in ZDEFs.