Fix: Forgot to set `monochrome` variable
In do_tint and do_monochrome, monochrome was not reset when you reset the screen
In do_colour, monochrome was not set despite being set, leading to losing the scratch palette.
Both of these issues have been corrected.
Also added, commented-out, a way to handle this which allows more direct control to the user; though it would require re-creating the presets, which will just be a bit of a chore.